more format string fixes

+++ init.c
@@ -216,7 +216,7 @@ void load_knowledge_list()
 		return;
 	}
 	char format[64];
-	safe_snprintf(format, sizeof(format), "%%hhd %%hd %%hhd %%hhd %%%u[^\n]", sizeof(knowledge_list->name) - 1);
+	safe_snprintf(format, sizeof(format), "%%hhd %%hd %%hhd %%hhd %%%d[^\n]", (int)sizeof(knowledge_list->name) - 1);
 	while (fgets(strLine, sizeof(strLine), f)) {
 		knowledge *k = knowledge_list + knowledge_count++;
 #ifdef ENGLISH
